requirements.txt
coverage
django>=1.10.3
djangorestframework>=3.5.2
markdown>=2.6.7
django-filter
django-nose
django-crispy-forms
django-gm2m
django-allauth
djangorestframework-csv
pika
git+https://github.com/Semprini/cbe.git
